pyproject.toml
[project]
name = "whatsappbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.30.0",
    "python-dotenv>=1.1.0",
    "whatsapp-api-client-python>=0.0.49",
]
